requirements.txt
# Frontend
streamlit>=1.28.0
streamlit-image-coordinates>=0.1.6

# Image processing
Pillow>=10.0.0

# Data handling
pyyaml>=6.0
pydantic>=2.0.0
numpy>=1.24.0

# API clients
openai>=1.0.0
replicate>=0.20.0
modal>=0.64.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
